vSPC: Handle VMOTION-ABORT.
VMOTION-ABORT is sent by the source host.  It does not require any
specific response or action, but we currently fail to recognize it,
which causes the _source_ vSPC connection to be terminated, and that
seems undesirable as the source should be fine -- it is the
_destination_ which has failed and is expected to disconnect.

This change handles it with a debug log message, but the side effect
of handling it is that the connection is no longer terminated.

Also fix the direction indicator for the debug message for
VMOTION-PEER-OK.

Change-Id: Id09ed363d6713c023323235a60d4c678c65079f1